Collection adjustments made and Player Inventory added.
A prompt now appears when the player enters a collectible object's range, and upon pressing 'E' the object will be added to their inventory, and then deleted from the scene. UI visuals have not been set up yet, and debug messages are still in the code to help with any bug testing.