Camera locking added
Some adjustments made to scene one, as well as a function added to the interactable object script that locks that camera when you're reading an objects text. For some reason though if you inspect the phone before grabbing the id, the id has the phone's text pop up. I have no idea why this is happening and I'll fix it later